+++ holberton.h
@@ -11,5 +11,5 @@ int _printf(const char *format, ...);
 int _strlen(char *);
 int _putchar(char c);
 int print_str(char *c);
-long int print_numbers(int n);
+int print_numbers(long int n);
 #endif
